V0.6.1.0 move external dlls to a solution folder move barcode specific controls to the barcode project sign main and charts assembly
