+++ CodeReason.Reports.Barcode/Document/TableCellBarcodeZenFramework.cs
@@ -1,0 +1,22 @@
+﻿/************************************************************************
+ * Copyright: Hans Wolff
+ *
+ * License:  This software abides by the LGPL license terms. For further
+ *           licensing information please see the top level LICENSE.txt 
+ *           file found in the root directory of CodeReason Reports.
+ *
+ * Author:   Hans Wolff
+ *
+ ************************************************************************/
+
+using CodeReason.Reports.Interfaces;
+
+namespace CodeReason.Reports.Barcode.Document
+{
+    /// <summary>
+    /// Table cell barcode ZenFramework
+    /// </summary>
+    public class TableCellBarcodeZenFramework : BarcodeZenFramework, ITableCellValue
+    {
+    }
+}
