+++ CodeReason.Reports.VS2010.sln
@@ -21,6 +21,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeReason.Reports.Math", "
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReportEditor", "ReportEditor\ReportEditor.csproj", "{63B7871A-0052-4401-8E4C-0072526364F9}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeReason.Reports.Barcode", "CodeReason.Reports.Barcode\CodeReason.Reports.Barcode.csproj", "{5F735C9E-B6C6-4424-A666-BBFAA73B4190}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -67,6 +69,10 @@ Global
 		{63B7871A-0052-4401-8E4C-0072526364F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{63B7871A-0052-4401-8E4C-0072526364F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{63B7871A-0052-4401-8E4C-0072526364F9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5F735C9E-B6C6-4424-A666-BBFAA73B4190}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5F735C9E-B6C6-4424-A666-BBFAA73B4190}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5F735C9E-B6C6-4424-A666-BBFAA73B4190}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5F735C9E-B6C6-4424-A666-BBFAA73B4190}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
