--- CodeReason.Reports/Document/Barcode/TableCellBarcodeZenFramework.cs
@@ -1,23 +1,0 @@
-﻿/************************************************************************
- * Copyright: Hans Wolff
- *
- * License:  This software abides by the LGPL license terms. For further
- *           licensing information please see the top level LICENSE.txt 
- *           file found in the root directory of CodeReason Reports.
- *
- * Author:   Hans Wolff
- *
- ************************************************************************/
-
-using CodeReason.Reports.Barcode;
-using CodeReason.Reports.Interfaces;
-
-namespace CodeReason.Reports.Document.Barcode
-{
-    /// <summary>
-    /// Table cell barcode ZenFramework
-    /// </summary>
-    public class TableCellBarcodeZenFramework : BarcodeZenFramework, ITableCellValue
-    {
-    }
-}
