+++ Samples/ChartReport/ChartReport.VS2010.csproj
@@ -113,13 +113,13 @@
     <AppDesigner Include="Properties\" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\CodeReason.Reports.Charts.Visifire\CodeReason.Reports.Charts.Visifire.VS2008.csproj">
-      <Project>{104E3C95-4F6C-4A1C-80B9-60B6BD784C51}</Project>
-      <Name>CodeReason.Reports.Charts.Visifire.VS2008</Name>
+    <ProjectReference Include="..\..\CodeReason.Reports.Charts.Visifire\CodeReason.Reports.Charts.Visifire.VS2010.csproj">
+      <Project>{104e3c95-4f6c-4a1c-80b9-60b6bd784c51}</Project>
+      <Name>CodeReason.Reports.Charts.Visifire.VS2010</Name>
     </ProjectReference>
-    <ProjectReference Include="..\..\CodeReason.Reports\CodeReason.Reports.VS2008.csproj">
-      <Project>{87A8D936-AE11-40AA-8927-A414C2225991}</Project>
-      <Name>CodeReason.Reports.VS2008</Name>
+    <ProjectReference Include="..\..\CodeReason.Reports\CodeReason.Reports.VS2010.csproj">
+      <Project>{87a8d936-ae11-40aa-8927-a414c2225991}</Project>
+      <Name>CodeReason.Reports.VS2010</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
