+++ Samples/DynamicReport/DynamicReport.VS2010.csproj
@@ -113,9 +113,9 @@
     <AppDesigner Include="Properties\" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\CodeReason.Reports\CodeReason.Reports.VS2008.csproj">
-      <Project>{87A8D936-AE11-40AA-8927-A414C2225991}</Project>
-      <Name>CodeReason.Reports.VS2008</Name>
+    <ProjectReference Include="..\..\CodeReason.Reports\CodeReason.Reports.VS2010.csproj">
+      <Project>{87a8d936-ae11-40aa-8927-a414c2225991}</Project>
+      <Name>CodeReason.Reports.VS2010</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
